fixing frontend

+++ client/src/store/actions/auth.js
@@ -1,41 +1,51 @@
-import * as api from  '../../api/index';
-import {AUTH,  LOGOUT, SHOWLOADER} from './actionTypes';
-import {dispatchFunction} from '../Utility';
+import * as api from '../../api/index'
+import { USER, LOGOUT, SHOWLOADER } from './actionTypes'
+import { loaderFunction } from '../Utility'
 
 
-
-export const signIn = (formData) =>  async (dispatch) => {
-   const data = await dispatchFunction(formData, api.signIn, dispatch);
-   data && dispatch({type: AUTH, payload: data.result})
-   data && localStorage.setItem('userToken', data.token);
+//if condition in every function because loaderFuntion will not return data if there's error
+export const signIn = (formData) => async (dispatch) => {
+  const data = await loaderFunction(formData, api.signIn, dispatch)
+  if (data) {
+    dispatch({ type: USER, payload: data.result })
+    localStorage.setItem('userToken', data.token)
+  }
 }
 
-export const signUp = (formData) => async(dispatch) => {
-    const data = await dispatchFunction(formData, api.signUp, dispatch);
-    data && dispatch({type: AUTH, payload: data.result});
-    data && localStorage.setItem('userToken', data.token);
+export const signUp = (formData) => async (dispatch) => {
+  const data = await loaderFunction(formData, api.signUp, dispatch)
+  if (data) {
+    dispatch({ type: USER, payload: data.result })
+    localStorage.setItem('userToken', data.token)
+  }
 }
 
 export const logOut = (history) => {
-    history.push('/');
-    localStorage.clear();
-    return{type:LOGOUT}
-} 
-
-export const getUser = (email) => async(dispatch) => {
-        const data = await dispatchFunction({email}, api.getUser, dispatch);
-        data && dispatch({type: AUTH, payload: data});
+  history.push('/')
+  localStorage.clear()
+  return { type: LOGOUT }
 }
 
-export const updateUser = (params) => async(dispatch) => {
-    const data = await dispatchFunction(params, api.updateUser, dispatch);
-    data && dispatch({type:SHOWLOADER, payload: data.message}); 
-    data && dispatch({type: AUTH, payload: data.result});
+export const getUser = (email) => async (dispatch) => {
+  const data = await loaderFunction({ email }, api.getUser, dispatch)
+  if (data) {
+    dispatch({ type: USER, payload: data.result })
+  }
 }
 
-export const deleteUser = (id, userType) => async(dispatch) => {
-    const {data} = await api.deleteUser(id,userType);
-    dispatch({type:SHOWLOADER, payload: data.message}); 
-    dispatch({type: AUTH, payload: ''});
-    localStorage.clear();
-} 
+export const updateUser = (params) => async (dispatch) => {
+  const data = await loaderFunction(params, api.updateUser, dispatch)
+  if (data) {
+    dispatch({ type: SHOWLOADER, payload: data.message })
+    dispatch({ type: USER, payload: data?.result })
+  }
+}
+
+export const deleteUser = (id, userType) => async (dispatch) => {
+  const { data } = await api.deleteUser(id, userType)
+  if (data) {
+    dispatch({ type: SHOWLOADER, payload: data.message })
+    dispatch({ type: USER, payload: '' })
+    localStorage.clear()
+  }
+}
